+++ Changelog.md
@@ -9,7 +9,7 @@ This tiny plug-in provides slideUp()/slideDown() like behavior for jquery
 using CSS transitions. These transitions tend to be very jerky on mobile
 so having a universal replacement is a common scenario.
 
-* **Add :containsNoCase and :startsWith jQuery Filters to ww.jquery
+* **Add :containsNoCase and :startsWith jQuery Filters to ww.jquery**
 <small>Westwind.Web/ww.jquery.js</small>
 Added these two filters that provide jQuery search filters. The former
 filter is especially useful to do dynamic page searches that show only